% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_allocation.R
\name{AllocateMarker}
\alias{AllocateMarker}
\alias{AllocateMarker.Seurat}
\alias{AllocateMarker.matrix}
\alias{AllocateMarker.Matrix}
\title{Assigns DEGs to the clusters as their markers}
\usage{
AllocateMarker(object, marker, ...)

\method{AllocateMarker}{Seurat}(
  object,
  marker,
  group_by = NULL,
  num_cores = 1,
  FDR_level = 0.05,
  ...
)

\method{AllocateMarker}{matrix}(object, marker, label, num_cores = 1, FDR_level = 0.05, ...)

\method{AllocateMarker}{Matrix}(object, marker, label, num_cores = 1, FDR_level = 0.05, ...)
}
\arguments{
\item{object}{A Seurat object or a matrix or dgCMatrix containing the normalized count matrix.}

\item{marker}{A vector containing names of marker genes, for example, output gene list from \code{\link{RunFestem}}.}

\item{...}{other parameters}

\item{group_by}{A character string or \code{NULL} (default). If \code{NULL}, \code{active.ident} is used
If a string, then the corresponding column in \code{meta.data} is taken as labels for cells.}

\item{num_cores}{Integer, number of CPU cores to use (default = 1).}

\item{FDR_level}{Numerical, nominal FDR level for multiple testing (default = 0.05).}

\item{label}{A vector containing labels for cells.}
}
\value{
A gene by cluster matrix indicating expression levels of each gene in different clusters. "A" stands for the highest expression level. We recommend using genes with expression level "A" and "B" as markers.
}
\description{
Given a set of markers (typically provided by \code{\link{RunFestem}}), 
use Scott-Knott test to decide in which groups they are up-regulated or down-regulated.
}
\examples{
# Load example data
data("example_data")
example_data <- Seurat::CreateSeuratObject(example_data$counts,meta.data = example_data$metadata)
example_data <- RunFestem(example_data,2,prior = "labels")
result <- AllocateMarker(example_data,rownames(example_data)[1:10],group_by = "labels")
print(result)
}
\references{
A. J. Scott, M. Knott, A cluster analysis method for grouping means in the analysis of variance. Biometrics 30, 507-512 (1974).
}
\seealso{
[RunFestem()]
}
