\name{EMDE-package}
\alias{sum_partial_r}
\alias{sum_partial_log_rr}
\alias{sum_partial_r_2}
\alias{sum_partial_log_rr_2}
\alias{EM_update_theta}
\alias{compute_w}
\alias{compute_alpha}
\alias{reduce_w}
\alias{compute_f}
\alias{compute_ff}
\alias{compute_pl}
\alias{compute_R1n}
\alias{FSEM}
\alias{homo_opt}
\docType{package}
\title{
\packageTitle{EMDE}
}
\description{
\packageDescription{EMDE}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{EMDE}
\packageIndices{EMDE}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{EMDE}

Maintainer: \packageMaintainer{EMDE}
}
\references{
No
}
\keyword{ package }
\seealso{
No
}
\examples{
EM_update_theta()
}
