% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{A toy example of scRNA-seq data in Seurat object}
\format{
A list containing expression counts and metadata of 100 genes across 485 cells. Can be convert to a Seurat object via \code{Seurat::CreateSeuratObject}
}
\usage{
example_data
}
\description{
A toy example of scRNA-seq data containing two cell types. Cell labels and counts are stored in Seurat object
}
\examples{
data("example_data")
example_data <- Seurat::CreateSeuratObject(example_data$counts,meta.data = example_data$metadata)
}
\keyword{datasets}
