% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_heatmap.R
\name{MarkerHeatmap}
\alias{MarkerHeatmap}
\alias{MarkerHeatmap.Seurat}
\alias{MarkerHeatmap.matrix}
\alias{MarkerHeatmap.Matrix}
\title{Heatmap visualization of markers identified by Festem}
\usage{
MarkerHeatmap(object, features, plot_cell_prop = 0.1, color_bar_max = 5, ...)

\method{MarkerHeatmap}{Seurat}(
  object,
  features,
  plot_cell_prop = 0.1,
  color_bar_max = 5,
  group_by = NULL,
  assay = "RNA",
  ...
)

\method{MarkerHeatmap}{matrix}(
  object,
  features,
  plot_cell_prop = 0.1,
  color_bar_max = 5,
  label,
  ...
)

\method{MarkerHeatmap}{Matrix}(
  object,
  features,
  plot_cell_prop = 0.1,
  color_bar_max = 5,
  label,
  ...
)
}
\arguments{
\item{object}{A Seurat object, matrix or dgCMatrix containing the original count matrix. 
If a Seurat object is provided, further information on cells can be included in the metadata slot 
and specified with parameters \code{group_by} 
If a matrix or dgCMatrix is provided, its rows should represent different
genes and columns stand for cells. We recommend setting the rownames as the name of genes.
If row names are not specified, numerical indices will be created to represent genes
according to their order in the input matrix.}

\item{features}{a vector of genes to plot.}

\item{plot_cell_prop}{A float specifying the proportion of cells to plot. Only has an effect when the total
cell number is larger than 500.}

\item{color_bar_max}{A float specifying the maximum range of the color map.}

\item{...}{other parameters}

\item{group_by}{A character string or \code{NULL} (default). If \code{NULL}, \code{active.ident} is used
If a string, then the corresponding column in \code{meta.data} is taken as labels for cells.}

\item{assay}{Name of Assay used}

\item{label}{A vector containing labels for cells.}
}
\description{
Heatmap of marker gene expressions with given group labels.
}
\examples{
# Load example data
data("example_data")
example_data <- Seurat::CreateSeuratObject(example_data$counts,meta.data = example_data$metadata)
MarkerHeatmap(example_data,features = rownames(example_data)[1:10],group_by = "labels")
}
