% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Festem-package.R
\docType{package}
\name{Festem}
\alias{Festem-package}
\alias{Festem}
\title{Festem: Feature Selection and Differential Expression Gene Analysis via EM-test}
\description{
Directly selecting differentially expressed genes for single-cell clustering analyses via EM-test. "Festem" also provides EM-test to test whether the number of component in mixtures of normal, Poisson or negative binomial distributions are larger than one. See Chen, Wang, et al (2023) <https://doi.org/10.1101/2023.07.26.550670> and Wang, Chen and Xi (2023) <https://doi.org/10.48550/arXiv.2306.12671> for more details.
}
\author{
\strong{Maintainer}: Zihao Chen \email{g.e.challenger@pku.edu.cn} (\href{https://orcid.org/0000-0002-1382-0409}{ORCID})

Authors:
\itemize{
  \item Changhu Wang \email{wangch156@pku.edu.cn} (\href{https://orcid.org/0009-0002-8567-0961}{ORCID})
  \item Ruibin Xi \email{ruibinxi@math.pku.edu.cn} (\href{https://orcid.org/0000-0001-7545-7361}{ORCID}) [copyright holder]
}

}
